+++ lib/src/cli/commands/commands.dart
@@ -1,6 +1,7 @@
 export 'cache_command.dart';
 export 'config_command.dart';
 export 'crawl_command.dart';
+export 'curate_command.dart';
 export 'cues_command.dart';
 export 'login_command.dart';
 export 'search_command.dart';