+++ lib/src/entities/entities.dart
@@ -1,4 +1,5 @@
 export 'cache_models.dart';
 export 'crawl_config.dart';
+export 'curate_config.dart';
 export 'reports/reports.dart';
 export 'sync_config.dart';